Merge pull request #436 from digitalinfinity:linux_rungcstress
Complete the port of GCStress to linux.
1. Fix bugs in implementation of BitTestAndSet etc.
2. Support passing in command line flags to GCStress
3. Use sigtrap for assertion failures to break into both lldb and gdb
4. Lower threshold of initial object count/op count to show progress
   on my VM. We'll have a future investigation item to understand what
   operations are slower on linux and how they can be sped up.

+++ lib/Common/Memory/Recycler.cpp
@@ -324,7 +324,7 @@ Recycler::Recycler(AllocationPolicyManager * policyManager, IdleDecommitPageAllo
 
 #ifdef ENABLE_DEBUG_CONFIG_OPTIONS
     // recycler requires at least Recycler::PrimaryMarkStackReservedPageCount to function properly for the main mark context
-    this->markContext.SetMaxPageCount(max(GetRecyclerFlagsTable().MaxMarkStackPageCount, Recycler::PrimaryMarkStackReservedPageCount));
+    this->markContext.SetMaxPageCount(max(static_cast<size_t>(GetRecyclerFlagsTable().MaxMarkStackPageCount), static_cast<size_t>(Recycler::PrimaryMarkStackReservedPageCount)));
     this->parallelMarkContext1.SetMaxPageCount(GetRecyclerFlagsTable().MaxMarkStackPageCount);
     this->parallelMarkContext2.SetMaxPageCount(GetRecyclerFlagsTable().MaxMarkStackPageCount);
     this->parallelMarkContext3.SetMaxPageCount(GetRecyclerFlagsTable().MaxMarkStackPageCount);
